reorganise folder structure for repositories
