+++ src/backend/Records/Persistence/Repositories/Database/PostgresUserRepository.cs
@@ -1,4 +1,5 @@
 using Application.Repositories;
+using Application.Repositories.Database;
 using Domain.Entities;
 using Microsoft.EntityFrameworkCore;
 using Persistence.Context;
